# Conflicts:
#	app/Console/Commands/InitWebhookDiscordChannels.php
#	app/Helpers/Discord/Features/RedeemCode.php

+++ app/Helpers/Discord/Features/RedeemCode.php
@@ -70,41 +70,38 @@ class RedeemCode extends DiscordMessage implements Feature
     public function createMainChannel(Guild $guild): void
     {
 
-        $category = Webhook::where('name', '=', 'Redeem Code Area')->first()->channel_id;
-        if ( !is_null($guild->channels->get('name', 'create-redeem-code'))) {
+        if (!is_null($guild->channels->get('name', 'create-redeem-code'))) {
 
             Webhook::updateOrCreate([
                 'name' => 'create-redeem-code'
             ], [
-                'name'       => $guild->channels->get('name', 'create-redeem-code')->name,
+                'name' => $guild->channels->get('name', 'create-redeem-code')->name,
                 'channel_id' => $guild->channels->get('name', 'create-redeem-code')->id,
-                'parent'     => false,
-                'parent_id'  => $category
-
+                'parent' => false
             ]);
 
             return;
         }
 
+        $category = Webhook::where('name', '=', 'Redeem Code Area')->first()->channel_id;
+
 
         $channel = $guild->channels->create([
-            'name'      => 'create-redeem-code',
-            'type'      => Channel::TYPE_GUILD_TEXT,
+            'name' => 'create-redeem-code',
+            'type' => Channel::TYPE_GUILD_TEXT,
             'parent_id' => $category,
-            'nsfw'      => false
+            'nsfw' => false
         ]);
 
 
-        $guild->channels->save($channel)->then(function (Channel $channel) use ($guild, $category) {
+        $guild->channels->save($channel)->then(function (Channel $channel) use ($guild) {
 
             Webhook::updateOrCreate([
                 'name' => $channel->name
             ], [
-                'name'       => $channel->name,
+                'name' => $channel->name,
                 'channel_id' => $channel->id,
-                'parent'     => false,
-                'parent_id'  => $category
-
+                'parent' => false
             ]);
 
             $this->createButtonChannel($guild);
@@ -118,7 +115,7 @@ class RedeemCode extends DiscordMessage implements Feature
     {
 
         $this->createCat();
-        $guild = $this->discord->guilds->get('id', env('DISCORD_BOT_GUILD_LOGS'));
+        $guild = $this->discord->guilds->get('id', env('DISCORD_BOT_GUILD'));
         $this->createMainChannel($guild);
         $this->createLogPage($guild);
 
@@ -145,15 +142,15 @@ class RedeemCode extends DiscordMessage implements Feature
     public function createCat(): Guild|string
     {
         try {
-            $guild = $this->discord->guilds->get('id', env('DISCORD_BOT_GUILD_LOGS'));
-            if ( !is_null($guild->channels->get('name', 'Redeem Code Area'))) {
+            $guild = $this->discord->guilds->get('id', env('DISCORD_BOT_GUILD'));
+            if (!is_null($guild->channels->get('name', 'Redeem Code Area'))) {
 
                 Webhook::updateOrCreate([
                     'name' => 'Redeem Code Area'
                 ], [
-                    'name'       => $guild->channels->get('name', 'Redeem Code Area')->name,
+                    'name' => $guild->channels->get('name', 'Redeem Code Area')->name,
                     'channel_id' => $guild->channels->get('name', 'Redeem Code Area')->id,
-                    'parent'     => true
+                    'parent' => true
                 ]);
 
                 return false;
@@ -167,9 +164,9 @@ class RedeemCode extends DiscordMessage implements Feature
                 Webhook::updateOrCreate([
                     'name' => $channel->name
                 ], [
-                    'name'       => $channel->name,
+                    'name' => $channel->name,
                     'channel_id' => $channel->id,
-                    'parent'     => true
+                    'parent' => true
                 ]);
 
                 return $guild;
@@ -184,41 +181,39 @@ class RedeemCode extends DiscordMessage implements Feature
     public function createLogPage(Guild $guild): void
     {
         try {
-            $category = Webhook::where('name', '=', 'DLPanel')->first()->channel_id;
+            $guild = $this->discord->guilds->get('id', env('DISCORD_BOT_GUILD_LOGS'));
 
-            if ( !is_null($guild->channels->get('name', 'dlp-redeem-codes'))) {
+            if (!is_null($guild->channels->get('name', 'redeem-codes'))) {
 
                 Webhook::updateOrCreate([
-                    'name' => 'dlp-redeem-codes'
+                    'name' => 'redeem-codes'
                 ], [
-                    'name'       => $guild->channels->get('name', 'dlp-redeem-codes')->name,
-                    'channel_id' => $guild->channels->get('name', 'dlp-redeem-codes')->id,
-                    'parent'     => false,
-                    'parent_id'  => $category
-
+                    'name' => $guild->channels->get('name', 'redeem-codes')->name,
+                    'channel_id' => $guild->channels->get('name', 'redeem-codes')->id,
+                    'parent' => false
                 ]);
 
                 return;
             }
 
+            $category = Webhook::where('name', '=', 'DLPanel')->first()->channel_id;
+
 
             $channel = $guild->channels->create([
-                'name'      => 'dlp-redeem-codes',
-                'type'      => Channel::TYPE_GUILD_TEXT,
+                'name' => 'redeem-codes',
+                'type' => Channel::TYPE_GUILD_TEXT,
                 'parent_id' => $category,
-                'nsfw'      => false
+                'nsfw' => false
             ]);
 
 
-            $guild->channels->save($channel)->then(function (Channel $channel) use ($guild, $category) {
+            $guild->channels->save($channel)->then(function (Channel $channel) use ($guild) {
                 Webhook::updateOrCreate([
                     'name' => $channel->name
                 ], [
-                    'name'       => $channel->name,
+                    'name' => $channel->name,
                     'channel_id' => $channel->id,
-                    'parent'     => false,
-                    'parent_id'  => $category
-
+                    'parent' => false
                 ]);
             })->done();
 
@@ -240,29 +235,29 @@ class RedeemCode extends DiscordMessage implements Feature
     {
         $redeem_code = $request->redeem_code()->first();
         $fields = [];
-        if ( !is_null($redeem_code)) {
+        if (!is_null($redeem_code)) {
             $fields[] = ['name' => 'Redeem Code:', 'value' => $redeem_code->code];
             $fields[] = ['name' => 'Created By:', 'value' => "<@{$request->request_by}>"];
             $fields[] = ['name' => 'Created Date:', 'value' => $redeem_code->created_at];
         }
         $fields[] = [
-            'name'  => 'Player Data:',
+            'name' => 'Player Data:',
             'value' => "**Discord** : <@{$request->discord_id}>\n**CitizenID** : {$request->citizenid}"
         ];
 
-        if ( !empty($request->vehicles)) {
+        if (!empty($request->vehicles)) {
             $fields[] = ['name' => 'Vehicles', 'value' => str_replace(',', "\n", $request->vehicles)];
         }
 
-        if ( !empty($request->weapons)) {
-            $fields[] = ['name' => 'Weapons', 'value' => str_replace(',', "\n", $request->weapons)];
+//        if (!empty($request->weapons)) {
+//            $fields[] = ['name' => 'Weapons', 'value' => str_replace(',', "\n", $request->weapons)];
+//        }
+
+        if (!empty($request->items)) {
+            $fields[] = ['name' => 'Items / Weapons', 'value' => str_replace(',', "\n", $request->items)];
         }
 
-        if ( !empty($request->items)) {
-            $fields[] = ['name' => 'Items', 'value' => str_replace(',', "\n", $request->items)];
-        }
-
-        if ( !empty($request->cash)) {
+        if (!empty($request->cash)) {
             $fields[] = ['name' => 'Cash', 'value' => $request->cash];
         }
 
@@ -281,7 +276,7 @@ class RedeemCode extends DiscordMessage implements Feature
             $member = $in->guild->members->get('id', $in->data->values[0]);
             $request = $this->getRequest(Player::getData($in->data->values[0]), $in->data->values[0]);
 
-            if ( !$request && !is_null($request)) {
+            if (!$request && !is_null($request)) {
                 $in->respondWithMessage(MessageBuilder::new()->setContent("<@{$member->user->id}> Doesn't have player in game."),
                     true);
                 $in->acknowledge();
@@ -291,7 +286,7 @@ class RedeemCode extends DiscordMessage implements Feature
             if (is_null($request)) {
                 $request = RedeemCodeRequest::create([
                     'discord_id' => $in->data->values[0],
-                    'citizenid'  => Player::getData($in->data->values[0])->citizenid,
+                    'citizenid' => Player::getData($in->data->values[0])->citizenid,
                     'request_by' => $in->user->id,
                     'created_at' => Carbon::now()
                 ]);
@@ -305,14 +300,14 @@ class RedeemCode extends DiscordMessage implements Feature
             $embed = $this->createSummaryRequest($request);
             $ar = $this->createButtons([
                 [
-                    'style'     => Button::STYLE_PRIMARY,
+                    'style' => Button::STYLE_PRIMARY,
                     'custom_id' => 'update_redeem+' . $member->user->id,
-                    'label'     => 'Update Redeem'
+                    'label' => 'Update Redeem'
                 ],
                 [
-                    'style'     => Button::STYLE_DANGER,
+                    'style' => Button::STYLE_DANGER,
                     'custom_id' => 'delete_redeem+' . $member->user->id,
-                    'label'     => 'Delete Redeem'
+                    'label' => 'Delete Redeem'
                 ]
             ]);
 
@@ -333,7 +328,7 @@ class RedeemCode extends DiscordMessage implements Feature
     {
         $builder = MessageBuilder::new();
         $select = StringSelect::new()->setPlaceholder('Select One Of The Choices');
-        $select->addOption(Option::new('Items', 'items'));
+        $select->addOption(Option::new('Items / Weapons', 'items'));
         $select->addOption(Option::new('Weapons', 'weapons'));
         $select->addOption(Option::new('Vehicles', 'vehicles'));
         $select->addOption(Option::new('Cash', 'cash'));
@@ -368,7 +363,7 @@ class RedeemCode extends DiscordMessage implements Feature
                             $count = 0;
                         }
                     }
-                    $items .= "**" . $item->label . "**  - " . $key . "\n";
+                    $items .= "**" . $item->label . "** , ";
                 }
             }
             $this->sendInsertRedeemType($i, $member);
@@ -448,26 +443,20 @@ class RedeemCode extends DiscordMessage implements Feature
         return true;
     }
 
-    private function generateCode($length = 35)
+    private function generateCode($length = 11)
     {
-        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ.-/\+';
+        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
         $charactersLength = strlen($characters);
         $randomString = '';
-        for ($i = 0 ; $i < $length ; $i++) {
+        for ($i = 0; $i < $length; $i++) {
 
-            if (strlen($randomString) % 6 == 0) {
-                $randomString .= '.';
-            }
-            if (strlen($randomString) % 10 == 0) {
-                $randomString .= '-/';
-            }
             $randomString .= $characters[random_int(0, $charactersLength - 1)];
         }
 
-        $code = \App\Models\RedeemCode::where('code', '=', $randomString)->first();
-        if ( !is_null($code)) {
-            $this->generateCode();
-        }
+//        $code = \App\Models\RedeemCode::where('code', '=', $randomString)->first();
+//        if (!is_null($code)) {
+//            $this->generateCode();
+//        }
 
         return $randomString;
     }
@@ -480,17 +469,31 @@ class RedeemCode extends DiscordMessage implements Feature
         $request = RedeemCodeRequest::where(['discord_id' => $member->user->id, 'citizenid' => $cid])->first();
 
         $redeem = \App\Models\RedeemCode::where('redeem_request', '=', $request->id)->first();
+        $code = $this->generateCode(3) . '-' . $this->generateCode(5) . '-' . $this->generateCode(3);
         if (is_null($redeem)) {
-            $code = $this->generateCode();
-            \App\Models\RedeemCode::create([
+            $redeem = \App\Models\RedeemCode::create([
                 'redeem_request' => $request->id,
-                'code'           => $code,
-                'created_at'     => Carbon::now()
+                'code' => $code,
+                'created_at' => Carbon::now()
             ]);
         }
 
+        RedeemCodeRequestHistory::updateOrCreate(
+            [
+                'code' => $redeem->code,
+                'discord_id' => $request->discord_id,
+                'citizenid' => $request->citizenid
+            ], [
+            'request_by' => $request->request_by,
+            'weapons' => $request->weapons,
+            'vehicles' => $request->vehicles,
+            'items' => $request->items,
+            'cash' => $request->cash,
+        ]);
+
         $embed = $this->createSummaryRequest($request);
         $in->sendFollowUpMessage(MessageBuilder::new()->addEmbed($embed), true)->done();
+        $in->deleteOriginalResponse();
         $in->acknowledge();
 
         $webhook = Webhook::where('name', '=', 'redeem-codes')->first()->channel_id;
@@ -515,18 +518,7 @@ class RedeemCode extends DiscordMessage implements Feature
         $redeem = \App\Models\RedeemCode::where('redeem_request', '=', $request->id)->first();
 
         !is_null($redeem) && $redeem->delete();
-        RedeemCodeRequestHistory::create([
-            'request_by' => $request->request_by,
-            'discord_id' => $request->discord_id,
-            'citizenid'  => $request->citizenid,
-            'weapons'    => $request->weapons,
-            'vehicles'   => $request->vehicles,
-            'items'      => $request->items,
-            'code'       => $redeem?->code,
-            'cash'       => $request->cash,
-
-        ]);
-        $request->delete();
+        !is_null($request) && $request->delete();
 
         $embed = $this->embed($this->discord, [], 'Redeem Deleted');
         $interaction->message->delete();
